refactor logic in to multiple files
